refactor(fspl): rename fastlens package
