+++ docs/api/inverse_ray_lightcurve.html
@@ -53,6 +53,7 @@
 <ul class="current">
 <li class="toctree-l1 current"><a class="reference internal" href="../modules.html">API Reference</a><ul class="current">
 <li class="toctree-l2"><a class="reference internal" href="point_source.html">Point Source</a></li>
+<li class="toctree-l2"><a class="reference internal" href="fspl.html">Finite-source Point Lens</a></li>
 <li class="toctree-l2"><a class="reference internal" href="multipole.html">Multipole</a></li>
 <li class="toctree-l2 current"><a class="current reference internal" href="#">Inverse-Ray Lightcurve</a><ul>
 <li class="toctree-l3"><a class="reference internal" href="#microjax.inverse_ray.lightcurve.mag_binary"><code class="docutils literal notranslate"><span class="pre">mag_binary()</span></code></a></li>
