+++ docs/api/multipole.html
@@ -20,7 +20,7 @@
     <link rel="index" title="Index" href="../genindex.html" />
     <link rel="search" title="Search" href="../search.html" />
     <link rel="next" title="Inverse-Ray Lightcurve" href="inverse_ray_lightcurve.html" />
-    <link rel="prev" title="Point Source" href="point_source.html" /> 
+    <link rel="prev" title="Finite-source Point Lens" href="fspl.html" />
 </head>
 
 <body class="wy-body-for-nav"> 
@@ -53,6 +53,7 @@
 <ul class="current">
 <li class="toctree-l1 current"><a class="reference internal" href="../modules.html">API Reference</a><ul class="current">
 <li class="toctree-l2"><a class="reference internal" href="point_source.html">Point Source</a></li>
+<li class="toctree-l2"><a class="reference internal" href="fspl.html">Finite-source Point Lens</a></li>
 <li class="toctree-l2 current"><a class="current reference internal" href="#">Multipole</a><ul>
 <li class="toctree-l3"><a class="reference internal" href="#microjax.multipole.mag_hexadecapole"><code class="docutils literal notranslate"><span class="pre">mag_hexadecapole()</span></code></a></li>
 </ul>
@@ -133,7 +134,7 @@ corrections through the <code class="docutils literal notranslate"><span class="
            </div>
           </div>
           <footer><div class="rst-footer-buttons" role="navigation" aria-label="Footer">
-        <a href="point_source.html" class="btn btn-neutral float-left" title="Point Source" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left" aria-hidden="true"></span> Previous</a>
+        <a href="fspl.html" class="btn btn-neutral float-left" title="Finite-source Point Lens" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left" aria-hidden="true"></span> Previous</a>
         <a href="inverse_ray_lightcurve.html" class="btn btn-neutral float-right" title="Inverse-Ray Lightcurve" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right" aria-hidden="true"></span></a>
     </div>
