+++ docs/api/point_source.html
@@ -19,7 +19,7 @@
     <script src="../_static/js/theme.js"></script>
     <link rel="index" title="Index" href="../genindex.html" />
     <link rel="search" title="Search" href="../search.html" />
-    <link rel="next" title="Multipole" href="multipole.html" />
+    <link rel="next" title="Finite-source Point Lens" href="fspl.html" />
     <link rel="prev" title="API Reference" href="../modules.html" /> 
 </head>
 
@@ -61,6 +61,7 @@
 <li class="toctree-l3"><a class="reference internal" href="#microjax.point_source.mag_point_source"><code class="docutils literal notranslate"><span class="pre">mag_point_source()</span></code></a></li>
 </ul>
 </li>
+<li class="toctree-l2"><a class="reference internal" href="fspl.html">Finite-source Point Lens</a></li>
 <li class="toctree-l2"><a class="reference internal" href="multipole.html">Multipole</a></li>
 <li class="toctree-l2"><a class="reference internal" href="inverse_ray_lightcurve.html">Inverse-Ray Lightcurve</a></li>
 <li class="toctree-l2"><a class="reference internal" href="poly_solver.html">Polynomial Solver</a></li>
@@ -266,7 +267,7 @@ branches returned by <code class="docutils literal notranslate"><span class="pre
           </div>
           <footer><div class="rst-footer-buttons" role="navigation" aria-label="Footer">
         <a href="../modules.html" class="btn btn-neutral float-left" title="API Reference" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left" aria-hidden="true"></span> Previous</a>
-        <a href="multipole.html" class="btn btn-neutral float-right" title="Multipole" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right" aria-hidden="true"></span></a>
+        <a href="fspl.html" class="btn btn-neutral float-right" title="Finite-source Point Lens" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right" aria-hidden="true"></span></a>
     </div>
 
   <hr/>
