+++ docs/genindex.html
@@ -78,7 +78,9 @@
 <h1 id="index">Index</h1>
 
 <div class="genindex-jumpbox">
- <a href="#C"><strong>C</strong></a>
+ <a href="#A"><strong>A</strong></a>
+ | <a href="#B"><strong>B</strong></a>
+ | <a href="#C"><strong>C</strong></a>
  | <a href="#E"><strong>E</strong></a>
  | <a href="#F"><strong>F</strong></a>
  | <a href="#G"><strong>G</strong></a>
@@ -93,6 +95,33 @@
  | <a href="#V"><strong>V</strong></a>
  
 </div>
+<h2 id="A">A</h2>
+<table style="width: 100%" class="indextable genindextable"><tr>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="api/fspl.html#fspl_disk.A">A() (fspl_disk method)</a>
+
+      <ul>
+        <li><a href="api/fspl.html#fspl_ld1.A">(fspl_ld1 method)</a>
+</li>
+        <li><a href="api/fspl.html#fspl_ld2.A">(fspl_ld2 method)</a>
+</li>
+      </ul></li>
+  </ul></td>
+</tr></table>
+
+<h2 id="B">B</h2>
+<table style="width: 100%" class="indextable genindextable"><tr>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li>
+    built-in function
+
+      <ul>
+        <li><a href="api/fspl.html#fspl_point">fspl_point()</a>
+</li>
+      </ul></li>
+  </ul></td>
+</tr></table>
+
 <h2 id="C">C</h2>
 <table style="width: 100%" class="indextable genindextable"><tr>
   <td style="width: 33%; vertical-align: top;"><ul>
@@ -142,12 +171,25 @@
   <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="api/utils.html#microjax.utils.first_nonzero">first_nonzero() (in module microjax.utils)</a>
 </li>
-  </ul></td>
-  <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="api/utils.html#microjax.utils.first_zero">first_zero() (in module microjax.utils)</a>
 </li>
       <li><a href="api/trajectory.html#microjax.trajectory.HeliocentricEphemeris.from_horizons_vectors_table">from_horizons_vectors_table() (microjax.trajectory.HeliocentricEphemeris static method)</a>
 </li>
+      <li><a href="api/fspl.html#fspl_disk">fspl_disk (built-in class)</a>
+</li>
+  </ul></td>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="api/fspl.html#fspl_ld1">fspl_ld1 (built-in class)</a>
+</li>
+      <li><a href="api/fspl.html#fspl_ld2">fspl_ld2 (built-in class)</a>
+</li>
+      <li>
+    fspl_point()
+
+      <ul>
+        <li><a href="api/fspl.html#fspl_point">built-in function</a>
+</li>
+      </ul></li>
   </ul></td>
 </tr></table>
