+++ docs/index.html
@@ -146,6 +146,7 @@ on CPUs but are intended primarily for GPU execution.</p>
 <li class="toctree-l1"><a class="reference internal" href="usage.html">Usage Guide</a><ul>
 <li class="toctree-l2"><a class="reference internal" href="usage.html#common-setup">Common setup</a></li>
 <li class="toctree-l2"><a class="reference internal" href="usage.html#point-source-magnification">Point-source magnification</a></li>
+<li class="toctree-l2"><a class="reference internal" href="usage.html#finite-source-point-lens">Finite-source point lens</a></li>
 <li class="toctree-l2"><a class="reference internal" href="usage.html#finite-source-binary-lenses">Finite-source binary lenses</a></li>
 <li class="toctree-l2"><a class="reference internal" href="usage.html#triple-lenses">Triple lenses</a></li>
 <li class="toctree-l2"><a class="reference internal" href="usage.html#configuration">Configuration</a></li>
