+++ docs/modules.html
@@ -53,6 +53,7 @@
 <ul class="current">
 <li class="toctree-l1 current"><a class="current reference internal" href="#">API Reference</a><ul>
 <li class="toctree-l2"><a class="reference internal" href="api/point_source.html">Point Source</a></li>
+<li class="toctree-l2"><a class="reference internal" href="api/fspl.html">Finite-source Point Lens</a></li>
 <li class="toctree-l2"><a class="reference internal" href="api/multipole.html">Multipole</a></li>
 <li class="toctree-l2"><a class="reference internal" href="api/inverse_ray_lightcurve.html">Inverse-Ray Lightcurve</a></li>
 <li class="toctree-l2"><a class="reference internal" href="api/poly_solver.html">Polynomial Solver</a></li>
@@ -91,6 +92,7 @@
 <div class="toctree-wrapper compound">
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="api/point_source.html">Point Source</a></li>
+<li class="toctree-l1"><a class="reference internal" href="api/fspl.html">Finite-source Point Lens</a></li>
 <li class="toctree-l1"><a class="reference internal" href="api/multipole.html">Multipole</a></li>
 <li class="toctree-l1"><a class="reference internal" href="api/inverse_ray_lightcurve.html">Inverse-Ray Lightcurve</a></li>
 <li class="toctree-l1"><a class="reference internal" href="api/poly_solver.html">Polynomial Solver</a></li>
