+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({"alltitles":{"API Reference":[[8,null],[9,null]],"Accuracy and limitations":[[8,"accuracy-and-limitations"]],"Boundary integration in outline":[[11,"boundary-integration-in-outline"]],"Citing microJAX":[[8,"citing-microjax"]],"Common setup":[[11,"common-setup"]],"Configuration":[[11,"configuration"]],"Coordinate conventions":[[3,"coordinate-conventions"]],"Design highlights":[[4,"design-highlights"]],"Differentiation":[[11,"differentiation"]],"Failure behavior":[[11,"failure-behavior"]],"Finite-source binary lenses":[[11,"finite-source-binary-lenses"]],"Functions":[[3,"functions"]],"Getting Started":[[7,null]],"Gradient computations stall":[[10,"gradient-computations-stall"]],"Guides":[[8,null]],"Highlights":[[8,"highlights"]],"Import errors for optional dependencies":[[10,"import-errors-for-optional-dependencies"]],"Indices and tables":[[8,"indices-and-tables"]],"Installation":[[7,"installation"]],"Inverse-Ray Lightcurve":[[0,null]],"JAX cannot see my GPU":[[10,"jax-cannot-see-my-gpu"]],"Likelihood Utilities":[[1,null]],"Multipole":[[2,null]],"Notes on numerical behavior":[[4,"notes-on-numerical-behavior"]],"Performance and reproducibility":[[11,"performance-and-reproducibility"]],"Point Source":[[3,null]],"Point-source magnification":[[11,"point-source-magnification"]],"Polynomial Solver":[[4,null]],"Prerequisites":[[7,"prerequisites"]],"Quick peek":[[8,"quick-peek"]],"Release lineage":[[8,"release-lineage"]],"Still stuck?":[[10,"still-stuck"]],"Trajectories":[[5,null]],"Triple lenses":[[11,"triple-lenses"]],"Troubleshooting":[[10,null]],"Up next":[[7,"up-next"]],"Usage Guide":[[11,null]],"Utility Functions":[[6,null]],"Verify the environment":[[7,"verify-the-environment"]],"mag_binary is slow or runs out of memory":[[10,"mag-binary-is-slow-or-runs-out-of-memory"]],"microJAX":[[8,null]]},"docnames":["api/inverse_ray_lightcurve","api/likelihood","api/multipole","api/point_source","api/poly_solver","api/trajectory","api/utils","getting_started","index","modules","troubleshooting","usage"],"envversion":{"sphinx":66,"sphinx.domains.c":3,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":9,"sphinx.domains.index":1,"sphinx.domains.javascript":3,"sphinx.domains.math":2,"sphinx.domains.python":4,"sphinx.domains.rst":2,"sphinx.domains.std":2},"filenames":["api/inverse_ray_lightcurve.rst","api/likelihood.rst","api/multipole.rst","api/point_source.rst","api/poly_solver.rst","api/trajectory.rst","api/utils.rst","getting_started.rst","index.rst","modules.rst","troubleshooting.rst","usage.rst"],"indexentries":{"circular_orbital_motion_3d() (in module microjax.trajectory)":[[5,"microjax.trajectory.circular_orbital_motion_3d",false]],"circular_orbital_motion_3d_jit() (in module microjax.trajectory)":[[5,"microjax.trajectory.circular_orbital_motion_3d_jit",false]],"circular_orbital_motion_3d_state() (in module microjax.trajectory)":[[5,"microjax.trajectory.circular_orbital_motion_3d_state",false]],"circular_orbital_motion_3d_state_jit() (in module microjax.trajectory)":[[5,"microjax.trajectory.circular_orbital_motion_3d_state_jit",false]],"compute_parallax() (in module microjax.trajectory)":[[5,"microjax.trajectory.compute_parallax",false]],"compute_parallax_ephem() (in module microjax.trajectory)":[[5,"microjax.trajectory.compute_parallax_ephem",false]],"critical_and_caustic_curves() (in module microjax.point_source)":[[3,"microjax.point_source.critical_and_caustic_curves",false]],"earth_orbital_parallax_offsets() (in module microjax.trajectory)":[[5,"microjax.trajectory.earth_orbital_parallax_offsets",false]],"earth_orbital_parallax_offsets_jit() (in module microjax.trajectory)":[[5,"microjax.trajectory.earth_orbital_parallax_offsets_jit",false]],"earthorbitalparallaxprojector (class in microjax.trajectory)":[[5,"microjax.trajectory.EarthOrbitalParallaxProjector",false]],"elliptic_orbital_motion_3d() (in module microjax.trajectory)":[[5,"microjax.trajectory.elliptic_orbital_motion_3d",false]],"elliptic_orbital_motion_3d_jit() (in module microjax.trajectory)":[[5,"microjax.trajectory.elliptic_orbital_motion_3d_jit",false]],"elliptic_orbital_motion_3d_state() (in module microjax.trajectory)":[[5,"microjax.trajectory.elliptic_orbital_motion_3d_state",false]],"elliptic_orbital_motion_3d_state_jit() (in module microjax.trajectory)":[[5,"microjax.trajectory.elliptic_orbital_motion_3d_state_jit",false]],"first_nonzero() (in module microjax.utils)":[[6,"microjax.utils.first_nonzero",false]],"first_zero() (in module microjax.utils)":[[6,"microjax.utils.first_zero",false]],"from_horizons_vectors_table() (microjax.trajectory.heliocentricephemeris static method)":[[5,"microjax.trajectory.HeliocentricEphemeris.from_horizons_vectors_table",false]],"getpsi() (in module microjax.trajectory)":[[5,"microjax.trajectory.getpsi",false]],"heliocentricephemeris (class in microjax.trajectory)":[[5,"microjax.trajectory.HeliocentricEphemeris",false]],"last_nonzero() (in module microjax.utils)":[[6,"microjax.utils.last_nonzero",false]],"lens_eq() (in module microjax.point_source)":[[3,"microjax.point_source.lens_eq",false]],"lens_eq_det_jac() (in module microjax.point_source)":[[3,"microjax.point_source.lens_eq_det_jac",false]],"linear_chi2() (in module microjax.likelihood)":[[1,"microjax.likelihood.linear_chi2",false]],"linear_orbital_motion() (in module microjax.trajectory)":[[5,"microjax.trajectory.linear_orbital_motion",false]],"linear_orbital_motion_jit() (in module microjax.trajectory)":[[5,"microjax.trajectory.linear_orbital_motion_jit",false]],"linear_orbital_motion_state() (in module microjax.trajectory)":[[5,"microjax.trajectory.linear_orbital_motion_state",false]],"linear_orbital_motion_state_jit() (in module microjax.trajectory)":[[5,"microjax.trajectory.linear_orbital_motion_state_jit",false]],"load_builtin_earth_ephemeris() (in module microjax.trajectory)":[[5,"microjax.trajectory.load_builtin_earth_ephemeris",false]],"load_horizons_vectors_file() (in module microjax.trajectory)":[[5,"microjax.trajectory.load_horizons_vectors_file",false]],"mag_binary() (in module microjax.inverse_ray.lightcurve)":[[0,"microjax.inverse_ray.lightcurve.mag_binary",false]],"mag_hexadecapole() (in module microjax.multipole)":[[2,"microjax.multipole.mag_hexadecapole",false]],"mag_point_source() (in module microjax.point_source)":[[3,"microjax.point_source.mag_point_source",false]],"mag_triple() (in module microjax.inverse_ray.lightcurve)":[[0,"microjax.inverse_ray.lightcurve.mag_triple",false]],"match_points() (in module microjax.utils)":[[6,"microjax.utils.match_points",false]],"max_zero_avoiding() (in module microjax.utils)":[[6,"microjax.utils.max_zero_avoiding",false]],"mean_zero_avoiding() (in module microjax.utils)":[[6,"microjax.utils.mean_zero_avoiding",false]],"microjax.inverse_ray.lightcurve":[[0,"module-microjax.inverse_ray.lightcurve",false]],"microjax.likelihood":[[1,"module-microjax.likelihood",false]],"microjax.multipole":[[2,"module-microjax.multipole",false]],"microjax.point_source":[[3,"module-microjax.point_source",false]],"microjax.poly_solver":[[4,"module-microjax.poly_solver",false]],"microjax.trajectory":[[5,"module-microjax.trajectory",false]],"microjax.utils":[[6,"module-microjax.utils",false]],"min_zero_avoiding() (in module microjax.utils)":[[6,"microjax.utils.min_zero_avoiding",false]],"module":[[0,"module-microjax.inverse_ray.lightcurve",false],[1,"module-microjax.likelihood",false],[2,"module-microjax.multipole",false],[3,"module-microjax.point_source",false],[4,"module-microjax.poly_solver",false],[5,"module-microjax.trajectory",false],[6,"module-microjax.utils",false]],"nll_ulens() (in module microjax.likelihood)":[[1,"microjax.likelihood.nll_ulens",false]],"nll_ulens_general() (in module microjax.likelihood)":[[1,"microjax.likelihood.nll_ulens_general",false]],"peri_vernal() (in module microjax.trajectory)":[[5,"microjax.trajectory.peri_vernal",false]],"poly_roots() (in module microjax.poly_solver)":[[4,"microjax.poly_solver.poly_roots",false]],"poly_roots_ea_multi() (in module microjax.poly_solver)":[[4,"microjax.poly_solver.poly_roots_EA_multi",false]],"poly_roots_self_inversive_fixed() (in module microjax.poly_solver)":[[4,"microjax.poly_solver.poly_roots_self_inversive_fixed",false]],"poly_roots_self_inversive_robust_fixed() (in module microjax.poly_solver)":[[4,"microjax.poly_solver.poly_roots_self_inversive_robust_fixed",false]],"prepare_projection_basis() (in module microjax.trajectory)":[[5,"microjax.trajectory.prepare_projection_basis",false]],"project_earth_position() (in module microjax.trajectory)":[[5,"microjax.trajectory.project_earth_position",false]],"r (microjax.trajectory.heliocentricephemeris attribute)":[[5,"microjax.trajectory.HeliocentricEphemeris.r",false]],"set_parallax() (in module microjax.trajectory)":[[5,"microjax.trajectory.set_parallax",false]],"set_parallax_ephem() (in module microjax.trajectory)":[[5,"microjax.trajectory.set_parallax_ephem",false]],"sparse_argsort() (in module microjax.utils)":[[6,"microjax.utils.sparse_argsort",false]],"t (microjax.trajectory.heliocentricephemeris attribute)":[[5,"microjax.trajectory.HeliocentricEphemeris.t",false]],"to_rotating_lens_frame() (in module microjax.trajectory)":[[5,"microjax.trajectory.to_rotating_lens_frame",false]],"trapz_zero_avoiding() (in module microjax.utils)":[[6,"microjax.utils.trapz_zero_avoiding",false]],"tree_flatten() (microjax.trajectory.earthorbitalparallaxprojector method)":[[5,"microjax.trajectory.EarthOrbitalParallaxProjector.tree_flatten",false]],"tree_flatten() (microjax.trajectory.heliocentricephemeris method)":[[5,"microjax.trajectory.HeliocentricEphemeris.tree_flatten",false]],"tree_unflatten() (microjax.trajectory.earthorbitalparallaxprojector class method)":[[5,"microjax.trajectory.EarthOrbitalParallaxProjector.tree_unflatten",false]],"tree_unflatten() (microjax.trajectory.heliocentricephemeris class method)":[[5,"microjax.trajectory.HeliocentricEphemeris.tree_unflatten",false]],"v (microjax.trajectory.heliocentricephemeris attribute)":[[5,"microjax.trajectory.HeliocentricEphemeris.v",false]]},"objects":{"microjax":[[1,0,0,"-","likelihood"],[2,0,0,"-","multipole"],[3,0,0,"-","point_source"],[4,0,0,"-","poly_solver"],[5,0,0,"-","trajectory"],[6,0,0,"-","utils"]],"microjax.inverse_ray":[[0,0,0,"-","lightcurve"]],"microjax.inverse_ray.lightcurve":[[0,1,1,"","mag_binary"],[0,1,1,"","mag_triple"]],"microjax.likelihood":[[1,1,1,"","linear_chi2"],[1,1,1,"","nll_ulens"],[1,1,1,"","nll_ulens_general"]],"microjax.multipole":[[2,1,1,"","mag_hexadecapole"]],"microjax.point_source":[[3,1,1,"","critical_and_caustic_curves"],[3,1,1,"","lens_eq"],[3,1,1,"","lens_eq_det_jac"],[3,1,1,"","mag_point_source"]],"microjax.poly_solver":[[4,1,1,"","poly_roots"],[4,1,1,"","poly_roots_EA_multi"],[4,1,1,"","poly_roots_self_inversive_fixed"],[4,1,1,"","poly_roots_self_inversive_robust_fixed"]],"microjax.trajectory":[[5,2,1,"","EarthOrbitalParallaxProjector"],[5,2,1,"","HeliocentricEphemeris"],[5,1,1,"","circular_orbital_motion_3d"],[5,1,1,"","circular_orbital_motion_3d_jit"],[5,1,1,"","circular_orbital_motion_3d_state"],[5,1,1,"","circular_orbital_motion_3d_state_jit"],[5,1,1,"","compute_parallax"],[5,1,1,"","compute_parallax_ephem"],[5,1,1,"","earth_orbital_parallax_offsets"],[5,1,1,"","earth_orbital_parallax_offsets_jit"],[5,1,1,"","elliptic_orbital_motion_3d"],[5,1,1,"","elliptic_orbital_motion_3d_jit"],[5,1,1,"","elliptic_orbital_motion_3d_state"],[5,1,1,"","elliptic_orbital_motion_3d_state_jit"],[5,1,1,"","getpsi"],[5,1,1,"","linear_orbital_motion"],[5,1,1,"","linear_orbital_motion_jit"],[5,1,1,"","linear_orbital_motion_state"],[5,1,1,"","linear_orbital_motion_state_jit"],[5,1,1,"","load_builtin_earth_ephemeris"],[5,1,1,"","load_horizons_vectors_file"],[5,1,1,"","peri_vernal"],[5,1,1,"","prepare_projection_basis"],[5,1,1,"","project_earth_position"],[5,1,1,"","set_parallax"],[5,1,1,"","set_parallax_ephem"],[5,1,1,"","to_rotating_lens_frame"]],"microjax.trajectory.EarthOrbitalParallaxProjector":[[5,3,1,"","tree_flatten"],[5,3,1,"","tree_unflatten"]],"microjax.trajectory.HeliocentricEphemeris":[[5,3,1,"","from_horizons_vectors_table"],[5,4,1,"","r"],[5,4,1,"","t"],[5,3,1,"","tree_flatten"],[5,3,1,"","tree_unflatten"],[5,4,1,"","v"]],"microjax.utils":[[6,1,1,"","first_nonzero"],[6,1,1,"","first_zero"],[6,1,1,"","last_nonzero"],[6,1,1,"","match_points"],[6,1,1,"","max_zero_avoiding"],[6,1,1,"","mean_zero_avoiding"],[6,1,1,"","min_zero_avoiding"],[6,1,1,"","sparse_argsort"],[6,1,1,"","trapz_zero_avoiding"]]},"objnames":{"0":["py","module","Python module"],"1":["py","function","Python function"],"2":["py","class","Python class"],"3":["py","method","Python method"],"4":["py","attribute","Python attribute"]},"objtypes":{"0":"py:module","1":"py:function","2":"py:class","3":"py:method","4":"py:attribute"},"terms":{"02j":11,"05j":11,"10j":11,"1d":5,"1e":[4,7],"1j":[5,7,8,11],"2_450_000":5,"2d":3,"3d":5,"5e":11,"A":[0,1,8,10,11],"All":[1,3],"At":11,"Both":[0,3],"Each":11,"For":[0,2,3,4,5,7,8,11],"Here":11,"If":[0,1,5,6,8],"It":[6,8],"On":10,"Or":7,"Such":11,"The":[0,1,2,4,5,6,7,8,10,11],"They":0,"This":[1,3,4,5,11],"Up":8,"We":10,"Where":[0,8],"__version__":[7,10],"_default_max_it":4,"_images_point_sourc":[2,3],"_images_point_source_sequenti":3,"aberth":4,"absolut":[2,4,5],"acceler":[7,8,11],"accept":[3,5],"access":10,"accur":11,"accuraci":11,"across":[3,6],"activ":8,"actual":8,"ad":4,"add":[5,11],"addit":[2,3,8,11],"adjac":11,"ae":1,"ae1005":8,"al":2,"algorithm":6,"align":5,"along":[6,11],"alpha":[5,8,11],"alpha0":5,"alpha_t":5,"also":[10,11],"although":11,"amount":[0,8],"amplitud":[2,5],"analysi":[8,11],"analyt":1,"anchor":5,"angl":[0,3,5,8,11],"angular":[0,8,11],"ani":[1,2],"annual":5,"anomali":5,"api":[7,11],"apj":8,"appli":[2,5],"appropri":[3,11],"approxim":[0,2,5,8,11],"ar":5,"arbitrari":1,"arc":11,"archiv":8,"argsort":6,"arithmet":3,"arrang":11,"array":[0,1,2,3,4,5,6,8,11],"ascens":5,"assign":6,"associ":8,"assum":1,"au":5,"au_c_day":5,"auto":5,"automat":[0,5,8,11],"aux":5,"avail":11,"avoid":[0,4,6,10],"away":0,"axi":[5,6],"b":[1,6],"back":4,"bartol":3,"base":[1,4,5],"baselin":5,"basi":5,"batch":[3,4,8,10],"becom":11,"befor":[8,10,11],"behavior":8,"benchmark":11,"best":1,"better":7,"beyond":6,"binari":[0,2,3,5,7,8],"binarylightcurvekepl":5,"binarylightcurveorbit":5,"binarymagconfig":[0,8,11],"bit":[7,11],"blend":1,"block":5,"block_until_readi":[10,11],"bodi":3,"bool":[4,5],"boolean":2,"bound":[0,8,11],"boundari":[0,8],"branch":[2,3,5,11],"bright":[0,8,11],"broadcast":[1,5],"build":[5,10],"built":8,"bundl":5,"c":[1,5,10],"calcul":[0,8,10,11],"calendar":5,"call":[8,10,11],"can":[2,4,5,7,10,11],"candid":6,"captur":1,"carri":0,"cartesian":5,"case":4,"cassan":2,"caustic":[0,3,8,11],"cd":7,"celesti":5,"cell":11,"center":3,"centr":[0,3],"chang":[10,11],"check":[10,11],"chi":1,"chi2":1,"children":5,"choos":11,"circl":[3,11],"circular":[0,5,8,11],"circular_orbital_motion_3d":5,"circular_orbital_motion_3d_jit":5,"circular_orbital_motion_3d_st":5,"circular_orbital_motion_3d_state_jit":5,"circumfer":[8,10,11],"class":5,"classmethod":5,"clean":7,"clone":7,"close":1,"closest":[5,6],"code":[10,11],"coeff":4,"coeffici":[0,2,3,4],"coeffs_matrix":4,"collaps":3,"column":[1,5],"com":7,"combin":11,"commit":[7,8,11],"common":[7,8,10],"compar":[5,11],"compat":[3,5,8],"compil":[0,8,10,11],"complet":[10,11],"complex":[0,2,3,4,5,6,11],"compon":5,"comput":[0,1,2,3,5,8,11],"compute_parallax":5,"compute_parallax_ephem":5,"condit":4,"config":[0,7,8,11],"configur":[0,2,3,8],"confirm":[7,10],"connect":11,"constant":1,"construct":[0,3,4,5,8,11],"contain":[3,11],"continu":11,"contribut":[2,11],"control":[5,10],"convent":5,"coordin":[0,2,5,11],"core":3,"correct":[2,5],"correspond":[0,3,5,8],"cos":[8,11],"cost":[10,11],"count":5,"covari":1,"cpu":[7,10],"cpus":[8,11],"creat":[5,8,11],"critic":3,"critical_and_caustic_curv":3,"cross":[8,11],"cuda":[7,10],"current":[5,8,11],"curv":[0,3,8,10,11],"custom":4,"custom_init":4,"custom_jvp":4,"d":5,"dalpha_dt":5,"darken":[0,2,8,11],"date":5,"day":5,"debug":10,"dec":5,"dec_deg":5,"declin":5,"def":11,"default":[0,1,2,3,4,5,6,10,11],"defin":[1,5],"definit":[1,5],"deg2rad":[8,11],"degre":5,"demo":10,"denomin":4,"depend":[3,5,8],"deriv":[4,5,8,10],"describ":7,"design":[0,1],"det":3,"detail":[2,7],"detect":7,"determin":[3,11],"determinist":6,"dev":7,"develop":8,"deviat":1,"devic":[7,10],"diag":1,"diagnost":2,"diagon":1,"dict":3,"differ":5,"differenc":5,"differenti":[0,1,4,5,8,10],"difficult":8,"dimension":6,"dimensionless":5,"direct":[5,8,11],"distanc":0,"divid":11,"dmu_dq":11,"document":10,"doe":[6,11],"doi":8,"done":1,"dot":5,"doubl":[8,10],"downstream":11,"driver":10,"ds":5,"ds_dt":5,"ds_z":5,"dt":5,"dtn":5,"dum":5,"e":[4,5,7],"e1":[2,3],"e2":[2,3],"ea":4,"earli":4,"earth":5,"earth_orbital_parallax_offset":5,"earth_orbital_parallax_offsets_jit":5,"earthorbitalparallaxprojector":5,"east":5,"ecc":5,"eccentr":5,"eclipt":5,"ehrlich":4,"einstein":[0,2,11],"either":[5,7],"element":[5,6],"elliptic_orbital_motion_3d":5,"elliptic_orbital_motion_3d_jit":5,"elliptic_orbital_motion_3d_st":5,"elliptic_orbital_motion_3d_state_jit":5,"els":1,"empir":5,"enabl":[7,8,11],"enclos":[0,8],"encod":3,"enough":8,"ensur":10,"entri":[1,5,6,7],"environ":[8,10],"eoe":5,"eph":5,"ephemeri":5,"epoch":5,"equal":1,"equat":[3,5,11],"equatori":5,"equinox":5,"equival":[5,6,10],"err":1,"err_i":1,"error":[0,1,8,11],"estim":[0,8,11],"et":2,"evalu":[1,2,3,5,11],"even":7,"everi":[4,7,11],"exact":[7,8],"exampl":[5,7,10,11],"execut":[8,11],"exp":5,"expect":[2,3,11],"expens":[0,8],"explicit":11,"expos":11,"express":1,"extend":3,"f":3,"failur":8,"fall":4,"fals":[4,5],"far":[0,8],"fast":[0,1,6,8,11],"fb":1,"feed":4,"field":5,"final":6,"find":[4,11],"finder":4,"fine":[10,11],"finit":[0,2,5,8,10],"first":[0,3,6,8,10,11],"first_nonzero":6,"first_zero":6,"fit":[1,10],"fix":[0,4,5,7,8,10],"float":[0,1,2,4,5],"floor":5,"flow":5,"flux":[1,11],"follow":[2,5,7,10,11],"forc":10,"form":[1,3,5,11],"forward":[0,4,8,10,11],"forward_model":11,"found":5,"fraction":[3,5],"frame":[3,5],"friend":[1,3,4,5],"from_horizons_vectors_t":5,"fs":1,"full":[0,1,5,8,11],"fulli":4,"function":[0,5,8,11],"g":4,"gamma":2,"gaug":2,"gaussian":1,"general":1,"generat":3,"geometri":[5,10],"get":[4,8],"getpsi":5,"git":[7,8],"github":[7,10],"given":[3,4],"gpu":[8,11],"gpus":11,"gradient":[1,3,4,8,11],"gravit":3,"greedi":6,"grid":[5,8],"guarante":[0,8,11],"guard":5,"guess":[4,5],"guid":7,"h":8,"half":3,"hand":5,"handl":[0,3],"happi":10,"hardwar":11,"heliocentr":5,"heliocentricephemeri":5,"help":10,"helper":[3,6],"hexadecapol":2,"hjd":5,"horizon":5,"https":7,"icr":5,"ignor":5,"ill":4,"imag":[0,2,3,6,8,10,11],"imaginari":[0,4,11],"impact":5,"implement":[1,2,3,4,5,7,8,11],"implicit":[4,10],"import":[7,8,11],"improv":[4,10],"includ":[6,8,10,11],"increas":8,"independ":[1,3,11],"index":[6,8],"indic":6,"infer":[1,5],"inform":10,"initi":[3,4,5],"initial_roots_matrix":4,"input":[0,1,3,5,6],"insid":11,"inspir":3,"instal":[8,10],"instantan":5,"insuffici":2,"int":[2,3,4,5,6],"integr":[0,1,6,8],"intend":[8,10,11],"intercept":1,"intern":[3,5],"interp_uniform_linear":5,"interpol":5,"introduc":11,"invers":[3,4,9],"inverse_ray":[0,8,11],"isol":0,"issu":10,"iter":[4,5,6],"j":3,"jacfwd":11,"jacobian":[3,8,11],"jax":[0,1,2,3,4,5,6,7,8,11],"jax_enable_x64":[7,8,10,11],"jax_platform":10,"jaxlib":[7,8,10,11],"jd":5,"jit":[3,5,8,10],"jnp":[1,4,7,8,11],"jpl":5,"julian":5,"kawahara":8,"keep":5,"kepler":5,"kepler_newton_it":5,"keplerian":5,"lambda":1,"larger":[5,11],"last":6,"last_nonzero":6,"later":10,"latest":7,"law":11,"lax":4,"least":1,"len":[0,2,3,5,6,7,8,11],"length":10,"lens":[0,2,3,8],"lens_eq":3,"lens_eq_det_jac":3,"level":[7,8,10],"librari":8,"licens":3,"lie":11,"light":[0,5,8,10,11],"light_time_it":5,"lightcurv":9,"like":5,"likelihood":[9,11],"limb":[0,2,8,11],"line":[5,7,8],"linear":[0,1,2,5,8,11],"linear_chi2":1,"linear_orbital_mot":5,"linear_orbital_motion_jit":5,"linear_orbital_motion_st":5,"linear_orbital_motion_state_jit":5,"linspac":[7,8,11],"list":7,"load":5,"load_builtin_earth_ephemeri":5,"load_horizons_vectors_fil":5,"local":5,"locat":2,"log":1,"long":10,"los":5,"loss":0,"lower":10,"lt":5,"m":[1,7,10],"m1":3,"m2":3,"m_fs":1,"mag_binari":[0,8,11],"mag_hexadecapol":2,"mag_point_sourc":[3,7,8,11],"mag_tripl":[0,8,11],"magnif":[0,2,3,7,8],"mahalanobi":1,"make":[1,4],"manual":10,"map":[1,3,5,11],"margin":1,"mask":2,"mass":[0,3,11],"match":[3,5,6,7,10],"match_point":6,"matplotlib":[7,10],"matrix":[1,5,7,10],"max_it":4,"max_zero_avoid":6,"maximum":[4,6],"may":[2,5,11],"mean":[1,5,6],"mean_zero_avoid":6,"measur":10,"memori":8,"method":[4,7,8],"microjax":[0,1,2,3,4,5,6,7,10,11],"microjaxx":7,"microlens":[1,3,7,8],"mid":[2,3],"midpoint":3,"min_zero_avoid":6,"minim":[1,10],"minimum":6,"miss":[6,10],"mit":3,"miyazaki":8,"mode":[0,7,8,11],"model":[1,5,8,11],"moder":5,"modifi":3,"modul":[1,3,4,6,8],"moment":2,"motion":[5,8],"mu":[1,2,11],"mu_finit":8,"mu_ld":11,"mu_point":8,"mu_tripl":11,"multi":10,"multipl":4,"multipol":9,"must":5,"n":[1,4,5],"n_branch":3,"n_imag":2,"n_limb":[0,8,10,11],"name":7,"nan":[0,6,8,11],"ndarray":[1,5],"near":[0,4],"nearest":[5,6],"need":[0,8,10],"negat":1,"neighbor":5,"neighbour":6,"new":10,"newer":7,"newton":[4,5],"next":8,"nlens":[2,3,7,8,11],"nll":1,"nll_ulen":1,"nll_ulens_gener":1,"nois":1,"non":[1,6,11],"none":[4,5],"normal":[0,11],"north":5,"note":[0,1,2,3,5],"npts":3,"number":[2,3,4,5,11],"numer":[0,7,8,10,11],"numpi":[7,8,10,11],"nvidia":10,"object":[5,11],"obliqu":5,"observ":[1,5],"obtain":[2,11],"occupi":11,"offici":[7,10],"offset":5,"omit":6,"one":[6,8,11],"onli":[7,10],"onto":5,"open":10,"oper":[1,3],"optim":[1,6],"option":[1,2,3,4,5,6,7,8],"orbit":[5,8],"order":[4,6],"orient":8,"origin":[3,5],"orthonorm":5,"otherwis":6,"outlin":8,"output":[2,3,4,5,10],"outsid":5,"overlap":11,"p":5,"packag":11,"page":8,"pair":5,"paper":[7,8],"parallax":[5,8],"parallax_param":5,"param":[2,3],"paramet":[0,1,2,3,4,5,6,8,11],"parameter":5,"pars":5,"part":[0,4,11],"pass":[4,5,10,11],"patch":8,"path":[1,5],"pattern":11,"per":[1,4,5],"perform":[1,3,8],"peri_vern":5,"perihelion":5,"period":5,"permut":6,"perpendicular":5,"phi":5,"photometri":1,"pick":6,"piee":5,"pien":5,"pip":[7,10],"pitfal":7,"place":11,"plan":7,"plane":[0,3,5,11],"platform":[7,8,10],"plot":[7,10],"point":[1,2,5,6,7,8,9],"point_sourc":[2,3,7,8,11],"poly_root":4,"poly_roots_ea":4,"poly_roots_ea_multi":4,"poly_roots_self_inversive_fix":4,"poly_roots_self_inversive_robust_fix":4,"poly_solv":4,"polyd":4,"polynomi":[3,9,10],"polyv":4,"posit":[0,1,3,5,6,8,11],"posterior":1,"pre":5,"precis":[8,10,11],"precomput":[4,5],"predictor":1,"prepar":7,"prepare_projection_basi":5,"prerequisit":8,"present":[4,6],"preserv":7,"previous":3,"primarili":[8,10],"print":[7,10],"prior":1,"problem":[6,10],"process":10,"profil":11,"project":[0,5],"project_earth_posit":5,"projector":5,"propag":3,"provid":[1,3,4,5,6,7,8],"psi":[0,3,5,11],"psi_offset":5,"public":[3,7,8,11],"publish":7,"pull":10,"pure":[4,5],"pypi":7,"python":[7,10],"q":[0,3,5,7,8,11],"q3":[0,3,11],"q_east":5,"q_north":5,"qne0":5,"quadrupol":2,"quantiti":[2,5],"quick":7,"r":5,"r3":[0,2,3,11],"ra":5,"ra_deg":5,"radial":11,"radian":[5,11],"radii":11,"radius":[0,2,3,11],"rang":11,"rank":1,"rapid":[10,11],"rate":5,"rather":7,"ratio":[0,3,5],"ray":9,"real":[0,3,4,11],"recommend":[0,7,11],"recompil":10,"record":[7,8,11],"redesign":[7,8],"reduc":10,"refer":[3,5,7],"regardless":4,"region":[0,8,11],"relat":[0,11],"releas":7,"reliabl":6,"relianc":4,"remain":[3,4,7,11],"remov":5,"repeat":[0,8,10],"replac":[1,6],"report":[8,10,11],"repres":5,"reproduc":[7,8,10],"requir":[3,11],"research":8,"residu":[2,5],"resolut":0,"respons":1,"result":[0,1,6,10,11],"return":[0,1,2,3,4,5,6,8,11],"reus":3,"revers":4,"rg":5,"rho":[0,2,8,11],"right":5,"robust":4,"rocm":[7,10],"root":[3,4],"roots_init":4,"rotat":5,"rotaxis_deg":5,"rout":11,"routin":[1,3],"rr":5,"rule":5,"run":[7,8,11],"runtim":10,"s":[0,3,5,7,8,11],"s0":5,"s_t":5,"s_z":5,"sampl":[0,3,5,6,7,8,11],"scalar":[1,3,5],"scipi":6,"script":10,"seaborn":[7,10],"search":[5,6,8],"second":0,"see":[7,8],"segment":6,"select":[2,5,11],"self":4,"sensit":10,"separ":[0,3,5,11],"sequenc":[3,11],"seri":8,"set":[0,3,6,8,10,11],"set_parallax":5,"set_parallax_ephem":5,"setup":8,"sever":[10,11],"shape":[1,2,3,4,5],"share":6,"shift":[3,5],"shorten":10,"shotamiyazaki94":7,"sider":5,"sight":5,"sigma2_f":1,"sigma2_fb":1,"sigma2_ob":1,"sign":5,"signal":5,"signatur":5,"simpl":1,"sin":[5,8,11],"singl":[5,10],"skip":[5,6],"sky":5,"slope":1,"slow":[4,8],"slower":10,"small":[0,1,4,5],"smaller":10,"smi":10,"smoke":7,"smooth":11,"snippet":[7,10],"soe":5,"softwar":8,"solut":1,"solv":[3,5,6],"solver":[7,8,9,10,11],"sort":6,"sourc":[0,1,2,5,6,7,8,9,10],"space":5,"sparse_argsort":6,"squar":1,"stabil":[4,7],"stack":[5,7,10],"stall":8,"standard":[1,5],"start":8,"state":[5,11],"static":5,"stay":11,"step":[4,5,6,11],"still":[8,11],"stop":4,"store":[3,5],"str":5,"strict":5,"stuck":8,"subdivis":11,"subinterv":11,"substanti":[7,8,10],"subtract":5,"suffici":[0,6],"suggest":10,"suitabl":1,"sum":[2,3],"sum_i":1,"summaris":10,"support":[0,3,4],"surfac":11,"switch":11,"symmetr":[1,5],"system":[2,3,8,10],"sz_t":5,"szs":5,"t":[1,5,8,11],"t0":11,"t0_par":5,"t_jdtdb":5,"tabl":5,"table_np":5,"tabul":5,"tail_idx":6,"take":10,"tangent":5,"target":5,"tau":[5,8,11],"tdb":5,"te":11,"temporarili":10,"term":[1,2,5],"test":7,"third":[0,3,11],"though":7,"three":[3,8,11],"threshold":4,"tilt":5,"time":[5,8,11],"times_are_absolut":5,"tini":4,"to_rotating_lens_fram":5,"togeth":[8,10],"tol":4,"toler":4,"total":[3,11],"tperi":5,"trace":[0,8,10],"track":[3,6],"trajectori":[0,8,9,11],"transform":8,"trapezoid":6,"trapz_zero_avoid":6,"treat":[1,6,8],"tree_flatten":5,"tree_unflatten":5,"tref":5,"tri":11,"tripl":[0,2,3,7,8],"triple_config":11,"triplemagconfig":[0,11],"troubleshoot":[7,8],"true":[4,5,7,8,11],"tupl":[1,2,3,5],"turn":10,"tvernal":5,"two":[0,1,3,6],"type":[0,1,3,4,5,6],"typic":2,"u":5,"u0":11,"u1":[0,2,11],"unchang":11,"uniform":[0,2,5,8],"unit":[0,2,3,5,11],"universal":11,"unspecifi":4,"unus":6,"unweight":1,"updat":[4,7,8,11],"usag":[7,8,10],"use":[0,2,3,4,5,7,8,10,11],"use_hjd":5,"user":10,"util":[3,4,5,8,9],"v":5,"v0":8,"valid":[0,2,3,8,10,11],"valu":[0,1,3,4,5,6,8,10,11],"variabl":10,"varianc":1,"variant":1,"vbbinarylens":5,"vector":[0,4,5],"veloc":5,"veri":10,"verifi":8,"vernal":5,"version":[4,7,8,10,11],"via":[2,4,5],"visual":7,"vmap":8,"vne0":5,"vx":5,"vy":5,"vz":5,"w":[3,5,7,8,11],"w1":5,"w2":5,"w3":5,"w_i":1,"w_k":2,"w_len":5,"w_point":0,"wait":10,"walk":7,"warm":11,"weight":1,"well":5,"wheel":7,"wherev":11,"while_loop":4,"within":11,"without":[0,5,8,11],"work":[0,5,7,8],"workflow":11,"x":[1,3,5,6,8,11],"x_i":1,"xla_python_client_prealloc":10,"y":[1,3,5,6,11],"y_i":1,"year":5,"yield":1,"z":[2,3,5],"z_ca":3,"z_cr":3,"z_eq":5,"z_mask":2,"zenodo":8,"zero":[0,1,4,6,11],"\u03b4beta":5,"\u03b4tau":5},"titles":["Inverse-Ray Lightcurve","Likelihood Utilities","Multipole","Point Source","Polynomial Solver","Trajectories","Utility Functions","Getting Started","microJAX","API Reference","Troubleshooting","Usage Guide"],"titleterms":{"Up":7,"accuraci":8,"api":[8,9],"behavior":[4,11],"binari":11,"boundari":11,"cite":8,"common":11,"comput":10,"configur":11,"convent":3,"coordin":3,"depend":10,"design":4,"differenti":11,"environ":7,"error":10,"failur":11,"finit":11,"function":[3,6],"get":7,"gpu":10,"gradient":10,"guid":[8,11],"highlight":[4,8],"import":10,"indic":8,"instal":7,"integr":11,"invers":0,"jax":10,"lens":11,"lightcurv":0,"likelihood":1,"limit":8,"lineag":8,"mag_binari":10,"magnif":11,"memori":10,"microjax":8,"multipol":2,"next":7,"note":4,"numer":4,"option":10,"outlin":11,"peek":8,"perform":11,"point":[3,11],"polynomi":4,"prerequisit":7,"quick":8,"ray":0,"refer":[8,9],"releas":8,"reproduc":11,"run":10,"see":10,"setup":11,"slow":10,"solver":4,"sourc":[3,11],"stall":10,"start":7,"still":10,"stuck":10,"tabl":8,"trajectori":5,"tripl":11,"troubleshoot":10,"usag":11,"util":[1,6],"verifi":7}})
+Search.setIndex({"alltitles":{"API Reference":[[9,null],[10,null]],"Accuracy and limitations":[[9,"accuracy-and-limitations"]],"Boundary integration in outline":[[12,"boundary-integration-in-outline"]],"Citing microJAX":[[9,"citing-microjax"]],"Common setup":[[12,"common-setup"]],"Configuration":[[12,"configuration"]],"Coordinate conventions":[[4,"coordinate-conventions"]],"Design highlights":[[5,"design-highlights"]],"Differentiation":[[12,"differentiation"]],"Failure behavior":[[12,"failure-behavior"]],"Finite-source Point Lens":[[0,null]],"Finite-source binary lenses":[[12,"finite-source-binary-lenses"]],"Finite-source point lens":[[12,"finite-source-point-lens"]],"Functions":[[4,"functions"]],"Getting Started":[[8,null]],"Gradient computations stall":[[11,"gradient-computations-stall"]],"Guides":[[9,null]],"Highlights":[[9,"highlights"]],"Import errors for optional dependencies":[[11,"import-errors-for-optional-dependencies"]],"Indices and tables":[[9,"indices-and-tables"]],"Installation":[[8,"installation"]],"Inverse-Ray Lightcurve":[[1,null]],"JAX cannot see my GPU":[[11,"jax-cannot-see-my-gpu"]],"Likelihood Utilities":[[2,null]],"Multipole":[[3,null]],"Notes on numerical behavior":[[5,"notes-on-numerical-behavior"]],"Performance and reproducibility":[[12,"performance-and-reproducibility"]],"Point Source":[[4,null]],"Point-source magnification":[[12,"point-source-magnification"]],"Polynomial Solver":[[5,null]],"Prerequisites":[[8,"prerequisites"]],"Quick peek":[[9,"quick-peek"]],"Release lineage":[[9,"release-lineage"]],"Still stuck?":[[11,"still-stuck"]],"Trajectories":[[6,null]],"Triple lenses":[[12,"triple-lenses"]],"Troubleshooting":[[11,null]],"Up next":[[8,"up-next"]],"Usage Guide":[[12,null]],"Utility Functions":[[7,null]],"Verify the environment":[[8,"verify-the-environment"]],"mag_binary is slow or runs out of memory":[[11,"mag-binary-is-slow-or-runs-out-of-memory"]],"microJAX":[[9,null]]},"docnames":["api/fspl","api/inverse_ray_lightcurve","api/likelihood","api/multipole","api/point_source","api/poly_solver","api/trajectory","api/utils","getting_started","index","modules","troubleshooting","usage"],"envversion":{"sphinx":66,"sphinx.domains.c":3,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":9,"sphinx.domains.index":1,"sphinx.domains.javascript":3,"sphinx.domains.math":2,"sphinx.domains.python":4,"sphinx.domains.rst":2,"sphinx.domains.std":2},"filenames":["api/fspl.rst","api/inverse_ray_lightcurve.rst","api/likelihood.rst","api/multipole.rst","api/point_source.rst","api/poly_solver.rst","api/trajectory.rst","api/utils.rst","getting_started.rst","index.rst","modules.rst","troubleshooting.rst","usage.rst"],"indexentries":{"a() (fspl_disk method)":[[0,"fspl_disk.A",false]],"a() (fspl_ld1 method)":[[0,"fspl_ld1.A",false]],"a() (fspl_ld2 method)":[[0,"fspl_ld2.A",false]],"built-in function":[[0,"fspl_point",false]],"circular_orbital_motion_3d() (in module microjax.trajectory)":[[6,"microjax.trajectory.circular_orbital_motion_3d",false]],"circular_orbital_motion_3d_jit() (in module microjax.trajectory)":[[6,"microjax.trajectory.circular_orbital_motion_3d_jit",false]],"circular_orbital_motion_3d_state() (in module microjax.trajectory)":[[6,"microjax.trajectory.circular_orbital_motion_3d_state",false]],"circular_orbital_motion_3d_state_jit() (in module microjax.trajectory)":[[6,"microjax.trajectory.circular_orbital_motion_3d_state_jit",false]],"compute_parallax() (in module microjax.trajectory)":[[6,"microjax.trajectory.compute_parallax",false]],"compute_parallax_ephem() (in module microjax.trajectory)":[[6,"microjax.trajectory.compute_parallax_ephem",false]],"critical_and_caustic_curves() (in module microjax.point_source)":[[4,"microjax.point_source.critical_and_caustic_curves",false]],"earth_orbital_parallax_offsets() (in module microjax.trajectory)":[[6,"microjax.trajectory.earth_orbital_parallax_offsets",false]],"earth_orbital_parallax_offsets_jit() (in module microjax.trajectory)":[[6,"microjax.trajectory.earth_orbital_parallax_offsets_jit",false]],"earthorbitalparallaxprojector (class in microjax.trajectory)":[[6,"microjax.trajectory.EarthOrbitalParallaxProjector",false]],"elliptic_orbital_motion_3d() (in module microjax.trajectory)":[[6,"microjax.trajectory.elliptic_orbital_motion_3d",false]],"elliptic_orbital_motion_3d_jit() (in module microjax.trajectory)":[[6,"microjax.trajectory.elliptic_orbital_motion_3d_jit",false]],"elliptic_orbital_motion_3d_state() (in module microjax.trajectory)":[[6,"microjax.trajectory.elliptic_orbital_motion_3d_state",false]],"elliptic_orbital_motion_3d_state_jit() (in module microjax.trajectory)":[[6,"microjax.trajectory.elliptic_orbital_motion_3d_state_jit",false]],"first_nonzero() (in module microjax.utils)":[[7,"microjax.utils.first_nonzero",false]],"first_zero() (in module microjax.utils)":[[7,"microjax.utils.first_zero",false]],"from_horizons_vectors_table() (microjax.trajectory.heliocentricephemeris static method)":[[6,"microjax.trajectory.HeliocentricEphemeris.from_horizons_vectors_table",false]],"fspl_disk (built-in class)":[[0,"fspl_disk",false]],"fspl_ld1 (built-in class)":[[0,"fspl_ld1",false]],"fspl_ld2 (built-in class)":[[0,"fspl_ld2",false]],"fspl_point()":[[0,"fspl_point",false]],"getpsi() (in module microjax.trajectory)":[[6,"microjax.trajectory.getpsi",false]],"heliocentricephemeris (class in microjax.trajectory)":[[6,"microjax.trajectory.HeliocentricEphemeris",false]],"last_nonzero() (in module microjax.utils)":[[7,"microjax.utils.last_nonzero",false]],"lens_eq() (in module microjax.point_source)":[[4,"microjax.point_source.lens_eq",false]],"lens_eq_det_jac() (in module microjax.point_source)":[[4,"microjax.point_source.lens_eq_det_jac",false]],"linear_chi2() (in module microjax.likelihood)":[[2,"microjax.likelihood.linear_chi2",false]],"linear_orbital_motion() (in module microjax.trajectory)":[[6,"microjax.trajectory.linear_orbital_motion",false]],"linear_orbital_motion_jit() (in module microjax.trajectory)":[[6,"microjax.trajectory.linear_orbital_motion_jit",false]],"linear_orbital_motion_state() (in module microjax.trajectory)":[[6,"microjax.trajectory.linear_orbital_motion_state",false]],"linear_orbital_motion_state_jit() (in module microjax.trajectory)":[[6,"microjax.trajectory.linear_orbital_motion_state_jit",false]],"load_builtin_earth_ephemeris() (in module microjax.trajectory)":[[6,"microjax.trajectory.load_builtin_earth_ephemeris",false]],"load_horizons_vectors_file() (in module microjax.trajectory)":[[6,"microjax.trajectory.load_horizons_vectors_file",false]],"mag_binary() (in module microjax.inverse_ray.lightcurve)":[[1,"microjax.inverse_ray.lightcurve.mag_binary",false]],"mag_hexadecapole() (in module microjax.multipole)":[[3,"microjax.multipole.mag_hexadecapole",false]],"mag_point_source() (in module microjax.point_source)":[[4,"microjax.point_source.mag_point_source",false]],"mag_triple() (in module microjax.inverse_ray.lightcurve)":[[1,"microjax.inverse_ray.lightcurve.mag_triple",false]],"match_points() (in module microjax.utils)":[[7,"microjax.utils.match_points",false]],"max_zero_avoiding() (in module microjax.utils)":[[7,"microjax.utils.max_zero_avoiding",false]],"mean_zero_avoiding() (in module microjax.utils)":[[7,"microjax.utils.mean_zero_avoiding",false]],"microjax.inverse_ray.lightcurve":[[1,"module-microjax.inverse_ray.lightcurve",false]],"microjax.likelihood":[[2,"module-microjax.likelihood",false]],"microjax.multipole":[[3,"module-microjax.multipole",false]],"microjax.point_source":[[4,"module-microjax.point_source",false]],"microjax.poly_solver":[[5,"module-microjax.poly_solver",false]],"microjax.trajectory":[[6,"module-microjax.trajectory",false]],"microjax.utils":[[7,"module-microjax.utils",false]],"min_zero_avoiding() (in module microjax.utils)":[[7,"microjax.utils.min_zero_avoiding",false]],"module":[[1,"module-microjax.inverse_ray.lightcurve",false],[2,"module-microjax.likelihood",false],[3,"module-microjax.multipole",false],[4,"module-microjax.point_source",false],[5,"module-microjax.poly_solver",false],[6,"module-microjax.trajectory",false],[7,"module-microjax.utils",false]],"nll_ulens() (in module microjax.likelihood)":[[2,"microjax.likelihood.nll_ulens",false]],"nll_ulens_general() (in module microjax.likelihood)":[[2,"microjax.likelihood.nll_ulens_general",false]],"peri_vernal() (in module microjax.trajectory)":[[6,"microjax.trajectory.peri_vernal",false]],"poly_roots() (in module microjax.poly_solver)":[[5,"microjax.poly_solver.poly_roots",false]],"poly_roots_ea_multi() (in module microjax.poly_solver)":[[5,"microjax.poly_solver.poly_roots_EA_multi",false]],"poly_roots_self_inversive_fixed() (in module microjax.poly_solver)":[[5,"microjax.poly_solver.poly_roots_self_inversive_fixed",false]],"poly_roots_self_inversive_robust_fixed() (in module microjax.poly_solver)":[[5,"microjax.poly_solver.poly_roots_self_inversive_robust_fixed",false]],"prepare_projection_basis() (in module microjax.trajectory)":[[6,"microjax.trajectory.prepare_projection_basis",false]],"project_earth_position() (in module microjax.trajectory)":[[6,"microjax.trajectory.project_earth_position",false]],"r (microjax.trajectory.heliocentricephemeris attribute)":[[6,"microjax.trajectory.HeliocentricEphemeris.r",false]],"set_parallax() (in module microjax.trajectory)":[[6,"microjax.trajectory.set_parallax",false]],"set_parallax_ephem() (in module microjax.trajectory)":[[6,"microjax.trajectory.set_parallax_ephem",false]],"sparse_argsort() (in module microjax.utils)":[[7,"microjax.utils.sparse_argsort",false]],"t (microjax.trajectory.heliocentricephemeris attribute)":[[6,"microjax.trajectory.HeliocentricEphemeris.t",false]],"to_rotating_lens_frame() (in module microjax.trajectory)":[[6,"microjax.trajectory.to_rotating_lens_frame",false]],"trapz_zero_avoiding() (in module microjax.utils)":[[7,"microjax.utils.trapz_zero_avoiding",false]],"tree_flatten() (microjax.trajectory.earthorbitalparallaxprojector method)":[[6,"microjax.trajectory.EarthOrbitalParallaxProjector.tree_flatten",false]],"tree_flatten() (microjax.trajectory.heliocentricephemeris method)":[[6,"microjax.trajectory.HeliocentricEphemeris.tree_flatten",false]],"tree_unflatten() (microjax.trajectory.earthorbitalparallaxprojector class method)":[[6,"microjax.trajectory.EarthOrbitalParallaxProjector.tree_unflatten",false]],"tree_unflatten() (microjax.trajectory.heliocentricephemeris class method)":[[6,"microjax.trajectory.HeliocentricEphemeris.tree_unflatten",false]],"v (microjax.trajectory.heliocentricephemeris attribute)":[[6,"microjax.trajectory.HeliocentricEphemeris.v",false]]},"objects":{"":[[0,0,1,"","fspl_disk"],[0,0,1,"","fspl_ld1"],[0,0,1,"","fspl_ld2"],[0,2,1,"","fspl_point"]],"fspl_disk":[[0,1,1,"","A"]],"fspl_ld1":[[0,1,1,"","A"]],"fspl_ld2":[[0,1,1,"","A"]],"microjax":[[2,3,0,"-","likelihood"],[3,3,0,"-","multipole"],[4,3,0,"-","point_source"],[5,3,0,"-","poly_solver"],[6,3,0,"-","trajectory"],[7,3,0,"-","utils"]],"microjax.inverse_ray":[[1,3,0,"-","lightcurve"]],"microjax.inverse_ray.lightcurve":[[1,2,1,"","mag_binary"],[1,2,1,"","mag_triple"]],"microjax.likelihood":[[2,2,1,"","linear_chi2"],[2,2,1,"","nll_ulens"],[2,2,1,"","nll_ulens_general"]],"microjax.multipole":[[3,2,1,"","mag_hexadecapole"]],"microjax.point_source":[[4,2,1,"","critical_and_caustic_curves"],[4,2,1,"","lens_eq"],[4,2,1,"","lens_eq_det_jac"],[4,2,1,"","mag_point_source"]],"microjax.poly_solver":[[5,2,1,"","poly_roots"],[5,2,1,"","poly_roots_EA_multi"],[5,2,1,"","poly_roots_self_inversive_fixed"],[5,2,1,"","poly_roots_self_inversive_robust_fixed"]],"microjax.trajectory":[[6,0,1,"","EarthOrbitalParallaxProjector"],[6,0,1,"","HeliocentricEphemeris"],[6,2,1,"","circular_orbital_motion_3d"],[6,2,1,"","circular_orbital_motion_3d_jit"],[6,2,1,"","circular_orbital_motion_3d_state"],[6,2,1,"","circular_orbital_motion_3d_state_jit"],[6,2,1,"","compute_parallax"],[6,2,1,"","compute_parallax_ephem"],[6,2,1,"","earth_orbital_parallax_offsets"],[6,2,1,"","earth_orbital_parallax_offsets_jit"],[6,2,1,"","elliptic_orbital_motion_3d"],[6,2,1,"","elliptic_orbital_motion_3d_jit"],[6,2,1,"","elliptic_orbital_motion_3d_state"],[6,2,1,"","elliptic_orbital_motion_3d_state_jit"],[6,2,1,"","getpsi"],[6,2,1,"","linear_orbital_motion"],[6,2,1,"","linear_orbital_motion_jit"],[6,2,1,"","linear_orbital_motion_state"],[6,2,1,"","linear_orbital_motion_state_jit"],[6,2,1,"","load_builtin_earth_ephemeris"],[6,2,1,"","load_horizons_vectors_file"],[6,2,1,"","peri_vernal"],[6,2,1,"","prepare_projection_basis"],[6,2,1,"","project_earth_position"],[6,2,1,"","set_parallax"],[6,2,1,"","set_parallax_ephem"],[6,2,1,"","to_rotating_lens_frame"]],"microjax.trajectory.EarthOrbitalParallaxProjector":[[6,1,1,"","tree_flatten"],[6,1,1,"","tree_unflatten"]],"microjax.trajectory.HeliocentricEphemeris":[[6,1,1,"","from_horizons_vectors_table"],[6,4,1,"","r"],[6,4,1,"","t"],[6,1,1,"","tree_flatten"],[6,1,1,"","tree_unflatten"],[6,4,1,"","v"]],"microjax.utils":[[7,2,1,"","first_nonzero"],[7,2,1,"","first_zero"],[7,2,1,"","last_nonzero"],[7,2,1,"","match_points"],[7,2,1,"","max_zero_avoiding"],[7,2,1,"","mean_zero_avoiding"],[7,2,1,"","min_zero_avoiding"],[7,2,1,"","sparse_argsort"],[7,2,1,"","trapz_zero_avoiding"]]},"objnames":{"0":["py","class","Python class"],"1":["py","method","Python method"],"2":["py","function","Python function"],"3":["py","module","Python module"],"4":["py","attribute","Python attribute"]},"objtypes":{"0":"py:class","1":"py:method","2":"py:function","3":"py:module","4":"py:attribute"},"terms":{"02j":12,"05j":12,"10j":12,"1d":6,"1e":[5,8],"1j":[6,8,9,12],"2_450_000":6,"2d":4,"3d":6,"5e":12,"A":[0,1,2,9,11,12],"All":[2,4],"At":12,"Both":[1,4],"Each":12,"For":[1,3,4,5,6,8,9,12],"Here":12,"If":[1,2,6,7,9],"It":[7,9],"On":11,"Or":8,"Such":12,"The":[0,1,2,3,5,6,7,8,9,11,12],"They":1,"This":[2,4,5,6,12],"Up":9,"We":11,"Where":[1,9],"__version__":[8,11],"_default_max_it":5,"_images_point_sourc":[3,4],"_images_point_source_sequenti":4,"a0":[],"a1":[0,12],"a_p":[],"aberth":5,"absolut":[3,5,6],"acceler":[8,9,12],"accept":[4,6],"access":11,"accur":12,"accuraci":12,"across":[4,7],"activ":9,"actual":9,"ad":5,"add":[6,12],"addit":[3,4,9,12],"adjac":12,"ae":2,"ae1005":9,"al":3,"algorithm":7,"align":6,"along":[7,12],"alpha":[6,9,12],"alpha0":6,"alpha_t":6,"also":[11,12],"although":12,"amount":[1,9],"amplitud":[3,6],"analysi":[9,12],"analyt":2,"anchor":6,"angl":[1,4,6,9,12],"angular":[1,9,12],"ani":[2,3],"annual":6,"anomali":6,"api":[8,12],"apj":9,"appli":[3,6],"appropri":[4,12],"approxim":[1,3,6,9,12],"ar":6,"arbitrari":2,"arc":12,"archiv":9,"argsort":7,"arithmet":4,"arrang":12,"array":[1,2,3,4,5,6,7,9,12],"ascens":6,"assign":7,"associ":9,"assum":2,"au":6,"au_c_day":6,"auto":6,"automat":[1,6,9,12],"aux":6,"avail":12,"avoid":[1,5,7,11],"away":1,"axi":[6,7],"b":[2,7],"back":5,"bartol":4,"base":[2,5,6],"baselin":6,"basi":6,"batch":[4,5,9,11],"becom":12,"befor":[9,11,12],"behavior":9,"benchmark":12,"best":2,"better":8,"beyond":7,"binari":[1,3,4,6,8,9],"binarylightcurvekepl":6,"binarylightcurveorbit":6,"binarymagconfig":[1,9,12],"bit":[8,12],"blend":2,"block":6,"block_until_readi":[11,12],"bodi":4,"bool":[5,6],"boolean":3,"bound":[1,9,12],"boundari":[1,9],"branch":[3,4,6,12],"bright":[0,1,9,12],"broadcast":[2,6],"build":[6,11],"built":9,"bundl":6,"c":[2,6,11],"calcul":[0,1,9,11,12],"calendar":6,"call":[0,9,11,12],"can":[3,5,6,8,11,12],"candid":7,"captur":2,"carri":1,"cartesian":6,"case":5,"cassan":3,"caustic":[1,4,9,12],"cd":8,"celesti":6,"cell":12,"center":4,"centr":[1,4],"central":[],"chang":[11,12],"check":[11,12],"chi":2,"chi2":2,"children":6,"choos":12,"circl":[4,12],"circular":[0,1,6,9,12],"circular_orbital_motion_3d":6,"circular_orbital_motion_3d_jit":6,"circular_orbital_motion_3d_st":6,"circular_orbital_motion_3d_state_jit":6,"circumfer":[9,11,12],"class":[0,6],"classmethod":6,"clean":8,"clone":8,"close":2,"closest":[6,7],"code":[11,12],"coeff":5,"coeffici":[0,1,3,4,5],"coeffs_matrix":5,"collaps":4,"column":[2,6],"com":8,"combin":12,"commit":[8,9,12],"common":[8,9,11],"compar":[6,12],"compat":[4,6,9],"compil":[1,9,11,12],"complet":[11,12],"complex":[1,3,4,5,6,7,12],"compon":6,"comput":[1,2,3,4,6,9,12],"compute_parallax":6,"compute_parallax_ephem":6,"condit":5,"config":[1,8,9,12],"configur":[1,3,4,9],"confirm":[8,11],"connect":12,"constant":2,"construct":[0,1,4,5,6,9,12],"contain":[4,12],"continu":12,"contribut":[3,12],"control":[6,11],"convent":6,"coordin":[1,3,6,12],"core":4,"correct":[3,6],"correspond":[1,4,6,9],"cos":[9,12],"cost":[11,12],"count":6,"counterpart":[],"covari":2,"cpu":[8,11],"cpus":[9,12],"creat":[6,9,12],"critic":4,"critical_and_caustic_curv":4,"cross":[9,12],"cuda":[8,11],"current":[6,9,12],"curv":[1,4,9,11,12],"custom":5,"custom_init":5,"custom_jvp":5,"d":6,"dalpha_dt":6,"darken":[0,1,3,9,12],"date":6,"day":6,"debug":11,"dec":6,"dec_deg":6,"declin":6,"def":12,"default":[1,2,3,4,5,6,7,11,12],"defin":[2,6],"definit":[2,6],"deg2rad":[9,12],"degre":6,"demo":11,"denomin":5,"depend":[4,6,9],"deriv":[5,6,9,11],"describ":8,"design":[1,2],"det":4,"detail":[3,8],"detect":8,"determin":[4,12],"determinist":7,"dev":8,"develop":9,"deviat":2,"devic":[8,11],"diag":2,"diagnost":3,"diagon":2,"dict":4,"differ":6,"differenc":6,"differenti":[1,2,5,6,9,11],"difficult":9,"dimension":7,"dimensionless":6,"direct":[6,9,12],"disk":[],"distanc":1,"divid":12,"dmu_dq":12,"document":11,"doe":[7,12],"doi":9,"done":2,"dot":6,"doubl":[9,11],"downstream":12,"driver":11,"ds":6,"ds_dt":6,"ds_z":6,"dt":6,"dtn":6,"dum":6,"e":[5,6,8],"e1":[3,4],"e2":[3,4],"ea":5,"earli":5,"earth":6,"earth_orbital_parallax_offset":6,"earth_orbital_parallax_offsets_jit":6,"earthorbitalparallaxprojector":6,"east":6,"ecc":6,"eccentr":6,"eclipt":6,"ehrlich":5,"einstein":[1,3,12],"either":[6,8],"element":[6,7],"elliptic_orbital_motion_3d":6,"elliptic_orbital_motion_3d_jit":6,"elliptic_orbital_motion_3d_st":6,"elliptic_orbital_motion_3d_state_jit":6,"els":2,"empir":6,"enabl":[8,9,12],"enclos":[1,9],"encod":4,"enough":9,"ensur":11,"entri":[2,6,7,8],"environ":[9,11],"eoe":6,"eph":6,"ephemeri":6,"epoch":6,"equal":2,"equat":[4,6,12],"equatori":6,"equinox":6,"equival":[6,7,11],"err":2,"err_i":2,"error":[1,2,9,12],"estim":[1,9,12],"et":3,"evalu":[0,2,3,4,6,12],"even":8,"everi":[5,8,12],"exact":[8,9],"exampl":[6,8,11,12],"execut":[9,12],"exp":6,"expect":[3,4,12],"expens":[1,9],"explicit":12,"expos":12,"express":2,"extend":4,"f":4,"failur":9,"fall":5,"fals":[5,6],"far":[1,9],"fast":[1,2,7,9,12],"fb":2,"feed":5,"fft_logumax":[],"fft_logumin":[],"field":6,"final":7,"find":[5,12],"finder":5,"fine":[11,12],"finit":[1,3,6,9,10,11],"first":[1,4,7,9,11,12],"first_nonzero":7,"first_zero":7,"fit":[2,11],"fix":[1,5,6,8,9,11],"float":[1,2,3,5,6],"floor":6,"flow":6,"flux":[2,12],"follow":[3,6,8,11,12],"forc":11,"form":[2,4,6,12],"forward":[1,5,9,11,12],"forward_model":12,"found":6,"fourier":[],"fraction":[4,6],"frame":[4,6],"friend":[2,4,5,6],"from_horizons_vectors_t":6,"fs":2,"fspl":[0,12],"fspl_disk":[0,12],"fspl_ld1":[0,12],"fspl_ld2":0,"fspl_point":0,"full":[1,2,6,9,12],"fulli":5,"function":[1,6,9,12],"g":5,"gamma":3,"gaug":3,"gaussian":2,"general":2,"generat":4,"geometri":[6,11],"get":[5,9],"getpsi":6,"git":[8,9],"github":[8,11],"given":[4,5],"gpu":[9,12],"gpus":12,"gradient":[2,4,5,9,12],"gravit":4,"greedi":7,"grid":[6,9],"guarante":[1,9,12],"guard":6,"guess":[5,6],"guid":8,"h":9,"half":4,"hand":6,"handl":[1,4],"happi":11,"hardwar":12,"heliocentr":6,"heliocentricephemeri":6,"help":11,"helper":[4,7],"hexadecapol":3,"hjd":6,"horizon":6,"https":8,"icr":6,"ignor":6,"ill":5,"imag":[1,3,4,7,9,11,12],"imaginari":[1,5,12],"impact":6,"implement":[0,2,3,4,5,6,8,9,12],"implicit":[5,11],"import":[8,9,12],"improv":[5,11],"includ":[7,9,11,12],"increas":9,"independ":[2,4,12],"index":[7,9],"indic":7,"infer":[2,6],"inform":11,"initi":[4,5,6],"initial_roots_matrix":5,"input":[1,2,4,6,7],"insid":12,"inspir":4,"instal":[9,11],"instantan":6,"insuffici":3,"int":[3,4,5,6,7],"integr":[1,2,7,9],"intend":[9,11,12],"intercept":2,"intern":[4,6],"interp_uniform_linear":6,"interpol":6,"introduc":12,"invers":[4,5,10],"inverse_ray":[1,9,12],"isol":1,"issu":11,"iter":[5,6,7],"j":4,"jacfwd":12,"jacobian":[4,9,12],"jax":[1,2,3,4,5,6,7,8,9,12],"jax_enable_x64":[8,9,11,12],"jax_platform":11,"jaxlib":[8,9,11,12],"jd":6,"jit":[4,6,9,11],"jnp":[2,5,8,9,12],"jpl":6,"julian":6,"k":[],"kawahara":9,"keep":6,"kepler":6,"kepler_newton_it":6,"keplerian":6,"kwarg":[],"lambda":2,"larger":[6,12],"last":7,"last_nonzero":7,"later":11,"latest":8,"law":12,"lax":5,"least":2,"len":[1,3,4,6,7,8,9,10],"length":11,"lens":[1,3,4,9],"lens_eq":4,"lens_eq_det_jac":4,"level":[8,9,11],"librari":9,"licens":4,"lie":12,"light":[1,6,9,11,12],"light_time_it":6,"lightcurv":10,"like":6,"likelihood":[10,12],"limb":[0,1,3,9,12],"line":[6,8,9],"linear":[0,1,2,3,6,9,12],"linear_chi2":2,"linear_orbital_mot":6,"linear_orbital_motion_jit":6,"linear_orbital_motion_st":6,"linear_orbital_motion_state_jit":6,"linspac":[8,9,12],"list":8,"load":6,"load_builtin_earth_ephemeri":6,"load_horizons_vectors_fil":6,"local":6,"locat":3,"log":2,"long":11,"los":6,"loss":1,"lower":11,"lt":6,"m":[2,8,11],"m1":4,"m2":4,"m_fs":2,"mag_binari":[1,9,12],"mag_fft_jax":[],"mag_hexadecapol":3,"mag_point_sourc":[4,8,9,12],"mag_tripl":[1,9,12],"magnif":[0,1,3,4,8,9],"magnifi":[0,12],"mahalanobi":2,"make":[2,5],"manual":11,"map":[2,4,6,12],"margin":2,"mask":3,"mass":[1,4,12],"match":[4,6,7,8,11],"match_point":7,"matplotlib":[8,11],"matrix":[2,6,8,11],"max_it":5,"max_zero_avoid":7,"maximum":[5,7],"may":[3,6,12],"mean":[2,6,7],"mean_zero_avoid":7,"measur":11,"memori":9,"method":[0,5,8,9,12],"microjax":[1,2,3,4,5,6,7,8,11,12],"microjaxx":8,"microlens":[2,4,8,9],"mid":[3,4],"midpoint":4,"min_zero_avoid":7,"minim":[2,11],"minimum":7,"miss":[7,11],"mit":4,"miyazaki":9,"mode":[1,8,9,12],"model":[2,6,9,12],"moder":6,"modifi":4,"modul":[2,4,5,7,9],"moment":3,"motion":[6,9],"mu":[2,3,12],"mu_finit":9,"mu_ld":12,"mu_limb_darken":12,"mu_point":9,"mu_tripl":12,"mu_uniform":12,"multi":11,"multipl":5,"multipol":10,"must":6,"n":[2,5,6],"n_branch":4,"n_fft":[],"n_imag":3,"n_limb":[1,9,11,12],"name":8,"nan":[1,7,9,12],"ndarray":[2,6],"near":[1,5],"nearest":[6,7],"need":[1,9,11],"negat":2,"neighbor":6,"neighbour":7,"new":11,"newer":8,"newton":[5,6],"next":9,"nlens":[3,4,8,9,12],"nll":2,"nll_ulen":2,"nll_ulens_gener":2,"nois":2,"non":[2,7,12],"none":[5,6],"normal":[1,12],"normalize_sk":[],"north":6,"note":[1,2,3,4,6],"npts":4,"number":[3,4,5,6,12],"numer":[1,8,9,11,12],"numpi":[8,9,11,12],"nvidia":11,"object":[6,12],"obliqu":6,"observ":[2,6],"obtain":[3,12],"occupi":12,"offici":[8,11],"offset":6,"omit":7,"onc":0,"one":[7,9,12],"onli":[8,11],"onto":6,"open":11,"oper":[2,4],"optim":[2,7],"option":[2,3,4,5,6,7,8,9],"orbit":[6,9],"order":[5,7],"orient":9,"origin":[4,6],"orthonorm":6,"otherwis":7,"outlin":9,"output":[3,4,5,6,11],"outsid":6,"overlap":12,"overrid":[],"p":6,"packag":12,"page":9,"pair":6,"paper":[8,9],"parallax":[6,9],"parallax_param":6,"param":[3,4],"paramet":[1,2,3,4,5,6,7,9,12],"parameter":6,"pars":6,"part":[1,5,12],"pass":[5,6,11,12],"patch":9,"path":[2,6],"pattern":12,"per":[2,5,6],"perform":[2,4,9],"peri_vern":6,"perihelion":6,"period":6,"permut":7,"perpendicular":6,"phi":6,"photometri":2,"pick":7,"piee":6,"pien":6,"pip":[8,11],"pipelin":[],"pitfal":8,"place":12,"plan":8,"plane":[1,4,6,12],"platform":[8,9,11],"plot":[8,11],"point":[2,3,6,7,8,9,10],"point_sourc":[3,4,8,9,12],"poly_root":5,"poly_roots_ea":5,"poly_roots_ea_multi":5,"poly_roots_self_inversive_fix":5,"poly_roots_self_inversive_robust_fix":5,"poly_solv":5,"polyd":5,"polynomi":[4,10,11],"polyv":5,"posit":[1,2,4,6,7,9,12],"posterior":2,"pre":6,"precis":[9,11,12],"precomput":[5,6],"predictor":2,"prepar":8,"prepare_projection_basi":6,"prerequisit":9,"present":[5,7],"preserv":8,"previous":4,"primarili":[9,11],"print":[8,11],"prior":2,"problem":[7,11],"process":11,"profil":[0,12],"project":[1,6],"project_earth_posit":6,"projector":6,"propag":4,"provid":[2,4,5,6,7,8,9],"psi":[1,4,6,12],"psi_offset":6,"public":[4,8,9,12],"publish":8,"pull":11,"pure":[5,6],"pypi":8,"python":[8,11],"q":[1,4,6,8,9,12],"q3":[1,4,12],"q_east":6,"q_north":6,"qne0":6,"quadrat":0,"quadrupol":3,"quantiti":[3,6],"quick":8,"r":6,"r3":[1,3,4,12],"ra":6,"ra_deg":6,"radial":[0,12],"radian":[6,12],"radii":12,"radius":[0,1,3,4,12],"rang":12,"rank":2,"rapid":[11,12],"rate":6,"rather":8,"ratio":[1,4,6],"ray":10,"real":[1,4,5,12],"recommend":[1,8,12],"recompil":11,"record":[8,9,12],"redesign":[8,9],"reduc":11,"refer":[4,6,8],"regardless":5,"region":[1,9,12],"relat":[1,12],"releas":8,"reliabl":7,"relianc":5,"remain":[4,5,8,12],"remov":6,"repeat":[1,9,11],"replac":[2,7],"report":[9,11,12],"repres":6,"reproduc":[8,9,11],"requir":[4,12],"research":9,"residu":[3,6],"resolut":1,"respons":2,"result":[1,2,7,11,12],"return":[1,2,3,4,5,6,7,9,12],"reus":4,"revers":5,"rg":6,"rho":[0,1,3,9,12],"rho_switch":[],"right":6,"robust":5,"rocm":[8,11],"root":[4,5],"roots_init":5,"rotat":6,"rotaxis_deg":6,"rout":12,"routin":[2,4],"rr":6,"rule":6,"run":[8,9,12],"runtim":11,"s":[1,4,6,8,9,12],"s0":6,"s_t":6,"s_z":6,"sampl":[1,4,6,7,8,9,12],"scalar":[2,4,6],"scipi":7,"script":11,"seaborn":[8,11],"search":[6,7,9],"second":1,"see":[8,9],"segment":7,"select":[3,6,12],"self":5,"sensit":11,"separ":[0,1,4,6,12],"sequenc":[4,12],"seri":9,"set":[1,4,7,9,11,12],"set_parallax":6,"set_parallax_ephem":6,"setup":9,"sever":[11,12],"shape":[2,3,4,5,6],"share":7,"shift":[4,6],"shorten":11,"shotamiyazaki94":8,"sider":6,"sight":6,"sigma2_f":2,"sigma2_fb":2,"sigma2_ob":2,"sign":6,"signal":6,"signatur":6,"simpl":2,"sin":[6,9,12],"singl":[0,6,11],"sk":[],"skip":[6,7],"sky":6,"slope":2,"slow":[5,9],"slower":11,"small":[1,2,5,6],"smaller":11,"smi":11,"smoke":8,"smooth":12,"snippet":[8,11],"soe":6,"softwar":9,"solut":2,"solv":[4,6,7],"solver":[0,8,9,10,11,12],"sort":7,"sourc":[1,2,3,6,7,8,9,10,11],"space":6,"sparse_argsort":7,"squar":2,"stabil":[5,8],"stack":[6,8,11],"stall":9,"standard":[2,6],"start":9,"state":[6,12],"static":6,"stay":12,"step":[5,6,7,12],"still":[9,12],"stop":5,"store":[4,6],"str":6,"strict":6,"stuck":9,"subclass":[],"subdivis":12,"subinterv":12,"substanti":[8,9,11],"subtract":6,"suffici":[1,7],"suggest":11,"suitabl":2,"sum":[3,4],"sum_i":2,"summaris":11,"support":[1,4,5],"surfac":12,"switch":12,"symmetr":[2,6],"system":[3,4,9,11],"sz_t":6,"szs":6,"t":[2,6,9,12],"t0":12,"t0_par":6,"t_jdtdb":6,"tabl":6,"table_np":6,"tabul":6,"tail_idx":7,"take":11,"tangent":6,"target":6,"tau":[6,9,12],"tdb":6,"te":12,"temporarili":11,"term":[2,3,6],"test":8,"third":[1,4,12],"though":8,"three":[4,9,12],"threshold":5,"tilt":6,"time":[6,9,12],"times_are_absolut":6,"tini":5,"to_rotating_lens_fram":6,"togeth":[9,11],"tol":5,"toler":5,"total":[4,12],"tperi":6,"trace":[1,9,11],"track":[4,7],"trajectori":[1,9,10,12],"transform":9,"trapezoid":7,"trapz_zero_avoid":7,"treat":[2,7,9],"tree_flatten":6,"tree_unflatten":6,"tref":6,"tri":12,"tripl":[1,3,4,8,9],"triple_config":12,"triplemagconfig":[1,12],"troubleshoot":[8,9],"true":[5,6,8,9,12],"tupl":[2,3,4,6],"turn":11,"tvernal":6,"two":[1,2,4,7],"type":[1,2,4,5,6,7],"typic":3,"u":[0,6,12],"u0":12,"u1":[1,3,12],"u_switch":[],"unchang":12,"unifi":[],"uniform":[0,1,3,6,9],"unit":[1,3,4,6,12],"universal":12,"unspecifi":5,"unus":7,"unweight":2,"updat":[5,8,9,12],"usag":[8,9,11],"use":[1,3,4,5,6,8,9,11,12],"use_hjd":6,"user":11,"util":[4,5,6,9,10],"v":6,"v0":9,"valid":[1,3,4,9,11,12],"valu":[1,2,4,5,6,7,9,11,12],"variabl":11,"varianc":2,"variant":2,"vbbinarylens":6,"vector":[1,5,6],"veloc":6,"veri":11,"verifi":9,"vernal":6,"version":[5,8,9,11,12],"via":[3,5,6],"visual":8,"vmap":9,"vne0":6,"vx":6,"vy":6,"vz":6,"w":[4,6,8,9,12],"w1":6,"w2":6,"w3":6,"w_i":2,"w_k":3,"w_len":6,"w_point":1,"wait":11,"walk":8,"warm":12,"weight":2,"well":6,"wheel":8,"wherev":12,"while_loop":5,"within":12,"without":[1,6,9,12],"work":[1,6,8,9],"workflow":12,"x":[2,4,6,7,9,12],"x_i":2,"xla_python_client_prealloc":11,"y":[2,4,6,7,12],"y_i":2,"year":6,"yield":2,"z":[3,4,6],"z_ca":4,"z_cr":4,"z_eq":6,"z_mask":3,"zenodo":9,"zero":[1,2,5,7,12],"\u03b4beta":6,"\u03b4tau":6,"\u03c1":[]},"titles":["Finite-source Point Lens","Inverse-Ray Lightcurve","Likelihood Utilities","Multipole","Point Source","Polynomial Solver","Trajectories","Utility Functions","Getting Started","microJAX","API Reference","Troubleshooting","Usage Guide"],"titleterms":{"Up":8,"accuraci":9,"api":[9,10],"behavior":[5,12],"binari":12,"boundari":12,"cite":9,"common":12,"comput":11,"configur":12,"convent":4,"coordin":4,"depend":11,"design":5,"differenti":12,"environ":8,"error":11,"failur":12,"finit":[0,12],"function":[4,7],"get":8,"gpu":11,"gradient":11,"guid":[9,12],"highlight":[5,9],"import":11,"indic":9,"instal":8,"integr":12,"invers":1,"jax":11,"len":[0,12],"lens":12,"lightcurv":1,"likelihood":2,"limit":9,"lineag":9,"mag_binari":11,"magnif":12,"memori":11,"microjax":9,"multipol":3,"next":8,"note":5,"numer":5,"option":11,"outlin":12,"peek":9,"perform":12,"point":[0,4,12],"polynomi":5,"prerequisit":8,"quick":9,"ray":1,"refer":[9,10],"releas":9,"reproduc":12,"run":11,"see":11,"setup":12,"slow":11,"solver":5,"sourc":[0,4,12],"stall":11,"start":8,"still":11,"stuck":11,"tabl":9,"trajectori":6,"tripl":12,"troubleshoot":11,"usag":12,"util":[2,7],"verifi":8}})
