+++ docs_src/modules.rst
@@ -5,6 +5,7 @@ API Reference
    :maxdepth: 1
 
    api/point_source
+   api/fspl
    api/multipole
    api/inverse_ray_lightcurve
    api/poly_solver
